+++ packages/tools/transpiler/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dlightjs/transpiler
 
+## 0.10.5
+
+### Patch Changes
+
+- Updated dependencies
+  - babel-preset-dlight@0.10.5
+
 ## 0.10.4
 
 ### Patch Changes